+++ src/black/lines.py
@@ -21,13 +21,8 @@ from black.mode import Mode, Preview
 from black.nodes import STANDALONE_COMMENT, TEST_DESCENDANTS
 from black.nodes import BRACKETS, OPENING_BRACKETS, CLOSING_BRACKETS
 from black.nodes import syms, whitespace, replace_child, child_towards
-from black.nodes import (
-    is_multiline_string,
-    is_import,
-    is_type_comment,
-    is_within_annotation,
-    is_one_sequence_between,
-)
+from black.nodes import is_multiline_string, is_import, is_type_comment
+from black.nodes import is_one_sequence_between
 
 # types
 T = TypeVar("T")
@@ -258,7 +253,8 @@ class Line:
     ) -> bool:
         """Return True if we have a magic trailing comma, that is when:
         - there's a trailing comma here
-        - it's not a one-tuple (or the equivalent type hint)
+        - it's not a one-tuple
+        - it's not a single-element subscript
         Additionally, if ensure_removable:
         - it's not from square bracket indexing
         """
@@ -274,17 +270,15 @@ class Line:
 
         if closing.type == token.RSQB:
             if (
-                Preview.one_tuple_type in self.mode
-                and is_within_annotation(closing)
+                Preview.one_element_subscript in self.mode
+                and closing.parent
+                and closing.parent.type == syms.trailer
                 and closing.opening_bracket
                 and is_one_sequence_between(
-                    closing.opening_bracket, closing, self.leaves
-                )
-                and closing.parent
-                and closing.parent.prev_sibling
-                and (
-                    list(closing.parent.prev_sibling.leaves())[-1].value
-                    in ("tuple", "Tuple")
+                    closing.opening_bracket,
+                    closing,
+                    self.leaves,
+                    brackets=(token.LSQB, token.RSQB),
                 )
             ):
                 return False